refactor: integrate expire table management into query snapshot and tree processes

+++ gallery-backend/src/public/query_snapshot/start_loop.rs
@@ -21,51 +21,47 @@ impl QuerySnapshot {
 
             let write_txn = self.in_disk.begin_write().unwrap();
 
-            /*  let list: Vec<String> = write_txn
-                .list_tables()
-                .unwrap()
-                .map(|x| x.name().to_string())
-                .collect();
-            println!("list is {:?}", list); */
-
             write_txn
                 .list_tables()
                 .unwrap()
                 .par_bridge()
                 .for_each(|table_handle| {
-                    let timestamp = table_handle.name().parse::<u64>().unwrap();
-                    println!("VERSION_COUNT.load(Ordering::Relaxed) is {}", VERSION_COUNT.load(Ordering::Relaxed));
-                    println!("expired(timestamp) is {}", expired(timestamp));
-                    println!("VERSION_COUNT.load(Ordering::Relaxed) > timestamp && expired(timestamp) is {}", VERSION_COUNT.load(Ordering::Relaxed) > timestamp && expired(timestamp));
-                    if VERSION_COUNT.load(Ordering::Relaxed) > timestamp && expired(timestamp) {
-                        // 1 hours in milliseconds
-                        let binding = timestamp.to_string();
-                        let table_definition: TableDefinition<u64, PrefetchReturn> =
-                            TableDefinition::new(&binding);
-                        let read_txn = self.in_disk.begin_read().unwrap();
-                        let table = read_txn.open_table(table_definition).unwrap();
-                        let tree_snapshot_delete_queue: Vec<_> = table
-                            .iter()
-                            .unwrap()
-                            .filter_map(|result| {
-                                let (_, value) = result.unwrap();
-                                let prefetch_return = value.value();
-                                prefetch_return.map(|prefetch| prefetch.timestamp)
-                            })
-                            .collect();
+                    if let Ok(timestamp) = table_handle.name().parse::<u64>() {
+                        if VERSION_COUNT.load(Ordering::Relaxed) > timestamp && expired(timestamp) {
+                            // 1 hours in milliseconds
+                            let binding = timestamp.to_string();
+                            let table_definition: TableDefinition<u64, PrefetchReturn> =
+                                TableDefinition::new(&binding);
+                            let read_txn = self.in_disk.begin_read().unwrap();
+                            let table = read_txn.open_table(table_definition).unwrap();
+                            let tree_snapshot_delete_queue: Vec<_> = table
+                                .iter()
+                                .unwrap()
+                                .filter_map(|result| {
+                                    let (_, value) = result.unwrap();
+                                    let prefetch_return = value.value();
+                                    prefetch_return.map(|prefetch| prefetch.timestamp)
+                                })
+                                .collect();
 
-                        match write_txn.delete_table(table_handle) {
-                            Ok(true) => {
-                                info!("Delete table: {:?}", timestamp);
-                                TREE_SNAPSHOT_DELETE_QUEUE_SENDER
-                                    .get()
-                                    .unwrap()
-                                    .send(tree_snapshot_delete_queue)
-                                    .unwrap();
-                            }
-                            Ok(false) => error!("Failed to delete table: {:?}", timestamp),
-                            Err(e) => {
-                                error!("Failed to delete table: {:?}, error: {:?}", timestamp, e)
+                            match write_txn.delete_table(table_handle) {
+                                Ok(true) => {
+                                    info!("Delete query cache table: {:?}", timestamp);
+                                    TREE_SNAPSHOT_DELETE_QUEUE_SENDER
+                                        .get()
+                                        .unwrap()
+                                        .send(tree_snapshot_delete_queue)
+                                        .unwrap();
+                                }
+                                Ok(false) => {
+                                    error!("Failed to delete query cache table: {:?}", timestamp)
+                                }
+                                Err(e) => {
+                                    error!(
+                                        "Failed to delete query cache table: {:?}, error: {:?}",
+                                        timestamp, e
+                                    )
+                                }
                             }
                         }
                     }
@@ -100,6 +96,7 @@ impl QuerySnapshot {
 
                                         table.insert(expression_hashed, ref_data).unwrap();
                                     }
+
                                     txn.commit().unwrap();
                                     info!(duration = &*format!("{:?}", timer_start.elapsed());
                                         "Write query cache into disk"
